requirements.txt
Django==3.2.6
djangorestframework==3.12.4
celery==5.1.2
redis==3.5.3
Faker==9.3.0